# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  backend_route-fix-backup-20250603_161334_handlers.ErrorResponse:
    properties:
      details: {}
      error:
        type: string
    type: object
...
  request.CreatePermissionRequest:
    properties:
      action:
        maxLength: 50
        minLength: 2
        type: string
      description:
        type: string
      name:
        maxLength: 100
        minLength: 3
        type: string
      resource:
        maxLength: 50
        minLength: 2
        type: string
      scope:
        type: string
    required:
    - action
    - name
    - resource
    - scope
    type: object
...
  request.UpdatePermissionRequest:
    properties:
      action:
        type: string
      description:
        type: string
      name:
        type: string
      resource:
        type: string
      scope:
        type: string
    type: object
...
  response.PaginationResponse:
    properties:
      current_page:
        type: integer
      from:
        type: integer
      last_page:
        type: integer
      per_page:
        type: integer
      to:
        type: integer
      total:
        type: integer
    type: object
...
  response.PermissionListResponse:
    properties:
      pagination:
        $ref: '#/definitions/response.PaginationResponse'
      permissions:
        items:
          $ref: '#/definitions/response.PermissionResponse'
        type: array
    type: object
  response.PermissionResponse:
    properties:
      action:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      resource:
        type: string
      scope:
        type: string
      updated_at:
        type: string
    type: object
...
host: reksolindo.opuschamber.com
info:
  contact: {}
  description: Multi-tenant enterprise asset management system
  title: Reksolindo Enterprise Asset Management API
  version: "1.0"
paths:
...
  /permissions:
    get:
      description: Retrieves a paginated list of permissions, optionally filtered.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: limit
        type: integer
      - description: Search term for permission name, resource, or action
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of permissions
          schema:
            $ref: '#/definitions/response.PermissionListResponse'
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List permissions
      tags:
      - permissions
    post:
      consumes:
      - application/json
      description: Creates a new permission.
      parameters:
      - description: Permission creation details
        in: body
        name: permission
        required: true
        schema:
          $ref: '#/definitions/request.CreatePermissionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Permission created successfully
          schema:
            $ref: '#/definitions/response.PermissionResponse'
        "400":
          description: Invalid request payload or validation error
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "403":
          description: Forbidden - Insufficient permissions
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "409":
          description: Permission name already exists
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new permission
      tags:
      - permissions
  /permissions/{id}:
    delete:
      description: Deletes a specific permission by its ID.
      parameters:
      - description: Permission ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Permission deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid Permission ID
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "404":
          description: Permission not found
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "409":
          description: Permission is in use and cannot be deleted
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete permission by ID
      tags:
      - permissions
    get:
      description: Retrieves details of a specific permission by its ID.
      parameters:
      - description: Permission ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Permission details
          schema:
            $ref: '#/definitions/response.PermissionResponse'
        "400":
          description: Invalid Permission ID
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "404":
          description: Permission not found
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get permission by ID
      tags:
      - permissions
    put:
      consumes:
      - application/json
      description: Updates details of a specific permission by its ID.
      parameters:
      - description: Permission ID
        in: path
        name: id
        required: true
        type: integer
      - description: Permission update details
        in: body
        name: permission
        required: true
        schema:
          $ref: '#/definitions/request.UpdatePermissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Permission updated successfully
          schema:
            $ref: '#/definitions/response.PermissionResponse'
        "400":
          description: Invalid Permission ID or request payload
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "404":
          description: Permission not found
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "409":
          description: Permission name already exists
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update permission by ID
      tags:
      - permissions
...
schemes:
- https
- http
swagger: "2.0"